+++ packages/core/vitest.config.ts
@@ -1,0 +1,3 @@
+import { configShared } from "../../vitest.shared";
+
+export default configShared;