+++ packages/next-plugin/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @kuma-ui/next-plugin
 
+## 0.5.1
+
+### Patch Changes
+
+- @kuma-ui/webpack-plugin@0.5.1
+
 ## 0.5.0
 
 ### Minor Changes